requirements.txt
python-docx>=1.1.2
reportlab>=4.2.2
